fixed autosave fixed autosaveinterval during race added text / button widgets added exit panel
git-svn-id: svn://84.194.112.164/bengine@366 2ee6282c-0497-4b59-8b7c-bcac2105c061

+++ src/scenes/entities/worldb.cpp
@@ -315,7 +315,7 @@ void WorldB::autosaveCritters()
 {
 	if ( *critter_autosaveinterval > 0 )
 	{
-		autosaveCounter += (Timer::Instance()->elapsed/1000);
+		autosaveCounter += ((float)Timer::Instance()->elapsed/1000);
 		if ( autosaveCounter > *critter_autosaveinterval )
 		{
 			autosaveCounter = 0.0f;
